requirements-dev.txt
caniusepython3
ipython
Click
jinja2
virtualenv
requests
honcho
psutil
python-crontab
semantic_version
GitPython
pytest
